Trigger a model firewall reconfigure when machines are created
Instead of creating when we first attempt to open a port

This is because models with no machines don't have model security
groups. They can be created when a new machine is, so there are two
events we may wish to configure the model security group
- On change to the config
- On machine creation/start

Change the machine watcher to watch machine agent start times, i.e.
when the machine is actually created

The only reason this worked prior was because the ec2 provider was not
returning the correct errortype, leading to the firewaller bouncing
until the sec group was created. Resolve this by returning a NotFound
error