Replace ubuntu with ubuntu-lite on CI tests when possible
This is part of an effort to ensure that our CI tests are more robust
and to avoid flaky tests. Ideally, all charms used in the integration
tests would be under our control, so this commit replaces ubuntu with
our ubuntu-lite when possible.
In some cases it's not possible because of ubuntu-lite not supporting
multiple bases or revisions.