Ensure the namespace for models
The following ensures the namespace (UUID) are allowed to be accessed.
This is the follow on work from PR #15569. By using the modelmanager
modle table from the controller we can see if any access can be
legitimately used. For controller namespaces no restriction is performed
and is allowed automatically. For everything else a look up on the
controller model table is required.

To make this more performant in production it might be wise to place
a cache in front of the db query. This should prevent requests for
the same namespace being accessed over and over when we actually know
it's fine. The cache can be inmemory and it may expire after a limited
time. We don't allow deletion of dbs currently, but it would be nice
to know when they have been removed.