Integration tests for GetDB
The following defines some integration tests for getting the db. This
includes a new dqliteAppIntegrationSuite. The new suite masks out the
creation of a new db for every test and closing the dbApp in the
FullSuite. The reason for this, is that the worker will manage the
lifecycle of the dqliteApp internally and the test suite calling
close more than once leads to a panic.

The rest is actually quite simple. Setup a new worker (now exported)
and run GetDB against the model list.

Additionally, we only create one worker per suite, which helps with
speed.