Assume manifest, fall back to series.
With the use of charmhub, series in a charm is ignored in favor of a
charm's manifest. As all charmhub released charms have a manifest, they
are essentially metadata v2 charm. This leads to different behavior
with deploying the same charm blog as a local charm, especially when
the series = kubernetes. Kubernetes is not a valid base from a manifest.
Seen in LP 1986382.

This change ensures juju prefers a manifest in any charm deployed, only
falling back to series from metadata if no manifest is available. Nor do
we add kubernetes as a series to the returned list.