Check operator status when filtering application in juju status
In caas models, applications can go into an error state not because of
the unit workload_statuses or the agent_status but the operator status
in application's DetailedStatus. Without the added check, the
filtering in the FullStatus doens't consider such applications among
the errored ones because the agent_status would be idle and the unit
workload_status would be in waiting.