Changestream terms
The following introduces changestream terms in to the changestream
worker package. There is still a lot to do, but this is the first
step towards that.

The concept of a term is to group a set of change events from the
change db as one. It's then possible to parallize all dispatches
of the change events to the subscribers. Once all events have been
read from the terms channel does a term become done. It's at that
point a new iteration on the changes db can be done.

If a slow read is being done from the term channel, the subscription
will be unsubscribed and it will be up to the watcher to resubscribe
by getting it's initial changes again.

There are a lot of moving parts here and this was spec'd out in
the Prague sprint. The changes to represent all that foundational
work.