Implement worker cache
The following adds a worker cache check before attempting to query the
actual db. This will be useful to prevent the thundering herd problem
of all facades requesting a model db.

This really simplifies the logic that was proposed previously with a
namespace cache. This way there is only ever one truth. The dbRunner
will always hold the valid db for the model.

If we ever need to remove the model we only need to do it in one location.