Allow closing of db worker gracefully
It's highly plausable that a db runner hasn't been started, for
example GetDB hasn't been called before close was called. This can
happen in a few scenarios. In the current implementation, no models
are calling db and in the future when this is supported, we might
attempt to close a db worker before it's actually started.

The solution here is to check for not found and return nil. That
way we're just gracefully allowing the closing of the db.