Wait for NICs to be up before restarting machine on spaces_ec2 tests
On spaces_ec2 tests, we first add a new nic for the current juju
machine, then we get it's ID from `ip link` and then we patch the
netplan config with this new NIC id and restart the machine.
This causes the test to fail sometimes because we get the NIC id before
it's UP and not shown in `ip link`. This patch adds a wait_for to make
sure that the NIC is UP before retrieving it's ID.