Introduction of a Term type
The term type defines a watermark of changes, with the explicit
condition that no other term is running whilst this is active. Marking
the term as done then allows another term to be processed. This makes
the event multiplexer coupled to the event stream, but does allow us
to step the event multiplexer if required.

Additionally, we can now drop the active channel in a subscription in
favour of the tomb.Dying channel. This simplification now means that
we have one source of truth for when a subscription becomes unsubscribed.
Cleaning this logic up makes it easier to grok about the lifecycle of
a subscription.