Add image-id constraint on AWS integration tests
This CI test will make sure that the correct AMI is selected when
adding a machine using the image-id constraint.
We test this by first adding creating a ec2 instance (ubuntu
jammy) that will be used as AMI builder. Then we create the ami
using the instanceId. Once the AMI is available, we add a juju machine
with the ami-id as image-id constraint, and at last we check that the
deployed machine has indeed the correct ami-id using the aws client.