Benchmark tests
The following adds some basic benchmarks to ensure that we're not
creating unbounded amount of allocations during a dispatch of a
subscription.

The benchmark is broken into two sections.

 1. Does how many events we send at once cause any issues?
 2. Does how many subscriptions we have cause any issues?

The way that Juju works is that we're morelikely to have a lot of
subscribers with bursts of events because of coalescing. The tests
are designed to send many events to ever increaing subscribers.

For non-matching requests (subscriptions that don't match the topic),
the benchmarks show that we're not creating a lot of garbage. For
when we do match, I believe we can spend some time optimising for a
better filtering logic. The process here is correctness first, and
then improve and optimise. It feels like we're creating a bit too
much garbage when dispatching a set of events.