Use ResolveForDeploy instead of ResolveWithPreferredChannel.
ResolveForDeploy is a single call to get all the data needed rather
than many calls. This is possible with the move to a server side
deploy method which does all the pieces together, rather than
individually.

Use the new BaseSelector as well, rather than moving to series only
to use the SeriesSelector.