Ensure we clean up subscriptions by ns
The following ensures that we clean up subscriptions by namespace.
As part of that work it helps to know the internal numbers for a given
event multiplexer. The common pattern for doing this is exposing a
Report method. As the event multiplexer is running in a lock less
fashion we have to ensure that we report in the right way. The
code does this by making a request to the event loop and handling
the response. We have a sane timeout for it it's not being processed
and just giving up.