Changestream watermark
The following defines a new change_log_witness table. This will allow
juju to see what change log ids have been read per agent. We will
eventually only remove the values that all controller agents have
witnessed (with caveats).

The changestream will attempt to write last_seen_at timestamp and
change_log_id every 5 seconds. We don't care if this fails to write
the watermark everytime, just that it writes it more often than not.

Note: if the system can't handle a write every 5 seconds for a
watermark, then we're in serious issues.

In a future PR, a change log pruner will come along and check the
witness table for all agents. If any agents are older than a given
timeframe (probably 10 minutes of not seen time), then that
information will be discarded. For everything that is in range, we'll
then get the minimum change_log_id and delete everything prior to that.
Optionally, it might be nice to also write out what we dropped to
a stash table or durable log for historical/audit reasons. It might
be wise to remove the pruned controller agents if they haven't been
seen for 20 minutes, as we're unlikely to see them again. If they
do come back, they'll make a new insert into the db as an upsert, so
it's not very likely that we care anyway.

Signed-off-by: Simon Richardson <stickupkid@gmail.com>