Stream Report
The defines a report method on the stream and the changestream worker.
This allows the understanding of what the progress is being made by
the changestream worker. The lastID is used by the stream to know the
current primary key being worked on. The recordedLastID is used to know
which lastID was actually sync'd to the db.

We expect the lastID and recordedLastID to be in close proximity. If
they're not, we can ascertain if there is a problem with the worker.
As worst best scenario is roughly 5 seconds (we have a timer that
beats every 5 seconds). The worst worst scenario is unbounded currently.

Signed-off-by: Simon Richardson <stickupkid@gmail.com>