Service should touch the db
If we remove the model uuid from the model_list, new and future requests
via GetDB should fail once we also stop the worker. Although it's called
DeleteDB, dqlite doesn't allow us to actually delete a db. There is a
not explaining this.

The service layer now directs the model_list removal to the state layer,
so it feels more complete.