Remove TxnNoRetry
There will never be a reason to not retry. It's much faster to retry
than to serialize to one db connection. The following changes are in
preparation for using sqlair as the default transaction runner.

Removing the TxnNoRetry requires the lease expiry to only be run on
the primary controller. Changing the manifold is only required to
make that happen.