Add sqlair to the dbaccessor
The following adds sqlair package to the juju project. In doing so it
pushes the sqlair DB into the dbaccessor. This is the lowest level that
we interact with a raw DB. In this way we never create sqlair wrappers
at the txn level. Instead we always have a sqlair DB and we just unwrap
it when required. This allows it to be a lot more cheaper in the usage.

For now we have Txn and StdTxn. Both run the same retry strategies, with
the same internals. The only difference is that Txn will use the new
shiny sqlair package and the StdTxn will use the stdlib one.

We will require another follow-up PR to move our domain logic over to
the Txn method instead of the StdTxn one. This will then allow us to
drop the custom mapping code we've been writing.

The code has to jump through some loop-holes around closing the db, as
that isn't directly exposed on the sqlair DB, but otherwise it's a
nice easy drop in replacement.