Limit concurrency
The following implements feedback from the benchmarking results
that @tlm has been investigating whilst scaling juju. In order
to prevent juju from thrashing the cpu with so many retries, we
instead use a semaphore to limit how many transactions can happen
at one particular time. This forces the backing off of transactions
at the source. If dqlite is struggling, there is no need to make it
worse. Instead, just force them to reign it back by blocking.

The changes here should not impact any small work load, so should be
fine to land. One improvement that might be worth considering in the
future, is to give certain namespaces different transaction runners,
for example, giving the controller namespace a larger chunck of
available slots. This will be easy to do, as we just push the txn
runner into the dbaccessor directly. That can be a future PR though.