[state] validate offer before running DB transaction
Adds logic to the validateOfferArgs method to check that the application and
endpoints exist in state.
No need for the explicit names.IsValidApplication check anymore, as this is
already done inside state.Application.

Also add a regression test and improve our test coverage here, to ensure we are
rejecting offers which reference invalid applications or endpoints.