Use correct application name in cliient for upload pending resources
Without this change the app deploys, but the local resources couldn't
be uploaded because the application name used is incorrect. This fixes
it to use the correct app name so the local resource uploads work just
fine.

It appears that we missed this one in the QAs for the local resource
PRs because we always tested it with an alias to make it easy for the
QAers (e.g. `j deploy juju-qa-test one ...`).