Fix setting the charmURL for subordinate units
https://bugs.launchpad.net/juju/+bug/1987332

So it looks like a while ago #14345 fixed the logic for processing
subordinate units by passing the unit's URL instead of the principle
application URL to correctly set the 'upgrading-from' value. As a
consequence of this, however, because of the logic that was already
there at the time, in the recursive call the unit's charm is the same
as the charm we're passing (as the 2nd parameter to processUnit),
which results in the charmURL not being set for subordinate units (so
therefore we don't see the charm value in the FullStatus in pylibjuju).

This fixes it by adding a boolean flag to the processUnit function to
bypass the existing logic for setting the charmURL in case we're
coming from the subordinate processing path. As a result the charm
field is correctly set for subordinate units in the FullStatus.