Support bases in bundles
Use bases throughout bundle deploy code, parsing immidiately from series
if necessary

This has meant changing over from SeriesSelector to BaseSelector, which
behaves slightly differently. Noteably, it will not select bionic, so a
lot of tests have had to change their base from bionic to focal

Unfortunately, the least bad option was to keep bases as strings in
Change structs. This means the marshalled nicely for Args as a string, and
omitempty is respected

Also, since GUIArgs output ordering is magic, we cannot replace series
with base. GUIArgs and the associated GetChanges facade endpoint should
be deprecated in favour of GetChangesMapArgs in Juju 4