Implement a Pinger Facade
The pinger timeout type is now renamed to pinger and move to core.
This should allow for unit testing of the pinger directly. With the
real facade interface we can then also directly test the facade easily.

Note: we've kept the pinger tests in the apiserver for now to ensure
that we've not broken anything.