Remove the legacy workers
The following removes the old legacy workers. This includes both
the strings and notify workers that depended directly on state. This
is interesting as the legacy workers are almost identical to the core
counter parts.

The strings worker was no longer used at all. As the min units worker
was the only one that was using it and event that worker wasn't being
used. So removing that dead branch was easy.

The notify worker was a bit harder, as the cert updater was using it
and that required a shim to move it over to the core watcher package.
Once that was done, it was then easy to remove the legacy package.