Remove fullDocumentBeforeChange field on mongodb changestream
Having the fullDocumentBeforeChange value set to 'off' in the mongodb
watcher results in this silent error:
```
BSON field '$changeStream.fullDocumentBeforeChange' is an unknown field
```
because this field is added on mongodb 6.0.

Since we are setting it to 'off' which is also the default value on
mongodb >= 6.0, it's safe to simply remove it.