Allow upgrades without specifying version
The following allow migrating to a new version of juju without
explicitly stating the newer major version (3.x -> 4.x). Juju
shouldn't have to know about new major versions, so to fix this,
we check that it's not in the version map, before checking if we're
moving to a larger major version.

This is required as 3.1 has been ear marked as the last LTS before
releasing 4.0. The target controller will then validate where the
model is coming from, and ensure it's with it's own range before
allowing it to migrate.