Fix the app name in the client side for the DeployFromRepository
This is a follow up/correction on
379591da8ca5bf9a37715ffcffd8a5447748937c, which incorrectly uses the
charm name for the application name. This was overlooked because those
names are often the same, however, they're not guaranteed to be the
same (e.g. `jameinel-ubuntu-lite` charm is for the application
`ubuntu-lite`).

So this change fixes it by distinguishing the charm name and the
application name, by using the `info` we get from the
DeployFromRepository endpoint, which has the actual app name that's
just deployed.