Add constraints functions to db errors
The following also adds the other valuable constraints checking
functions to the database errors package. There is no generic
generic IsErrConstraint check, as it is expected that you should
be checking for the specific constraint violation.

A DML statement can't possibly violate all/multiple constraints
at the same time, hence why it's not offered.

Additionally, we're leaning heavily on to the SQL here and ON
CONFLIC ... DO NOTHING should stay. SQL is the source.