Add apiserver helpers for watchers
We have a repeated pattern on the apiserver facades on which we
initialize a watcher, try to get it's `Changes()`, and if the returned
channel is closed, we fail with an error but before we check that the
watcher has been correctly killed, otherwise we return a specific error.
This pattern makes usage of the `state.EnsureErr()` function which is
problematic first because it belongs to the legacy mongo state, and
second because the watcher must implement the `Err()` function.

For new watchers we will not implement the `Errer` interface (meaning no
`Err()`) and so this new helper makes sure that the watcher (worker) is
correctly stopped in case the changes are closed, before registering
this watcher in the (api) resources.