Client now uses SupportedBases from ResolveCharm
This is a step towards replacing series with bases in juju

SupportedBases were previously added to the return of ResolveCharms
alongside SupportedSeries. The client now reads this instead, only
falling back for legacy support

ResolveCharms is used in many places, so a fair amount of conversion was
done. To keep the size down, in a few places we convert back to series
and maintain previous behaviour