Introduction of CODEOWNERS
The following is an experiment to introduce CODEOWNERS[1] that
allow to define owners to particular parts of the code. Whilst
developing a new feature it makes sense to put someone in charge
of a piece of code. Once the code has stabilized the removal of
the owners is acceptable. This should ensure that the feature
remains true to original design.

This shouldn't stop people changing the underlying code, all this
does ask an owner to review the proposed code.

 1. https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-code-owners#about-code-owners