Ensure that the table exists before pruning
The following just ensures that the table exists for that model in
the db, before attempting to prune it. This can happen if the model
is added to the model_list, yet the schema hasn't been applied. This
safety check just prevents lots of logs, it doesn't bring the pruner
down.