Allow migration from >=2.9.42 to 3.1.x
The following just allows clients to connect the 3.1.x apiserver.
Previously the code stated that if the minor version of the server
version was greater than 0, it couldn't allow a connection. Which
in term would return incompatible client error.

There is no point to this check. If the client is older, we should
allow it to query the min client versions. We should force the
client to jump to 3.0.x before 3.x.x. Removing the if statement and
adding a test fixes this.