Generic watchers
The following changes the core watcher to a generic watcher. This
includes changing the name to just Watcher. As this takes a type
we can now model all watchers.

Additionally, I've removed the intermidate type from the changes()
method. This has now become a type alias. It's possible to remove
this completely and just use bare types (<-chan T), but the amount
of code that has to change is a bit silly. For now I've created
a type alias for those legacy signatures. Everything should now
still compile and run, based on the type alias.