Remove string resource type
The following removes the string resource type from the apiserver
common package. The string resource was used to thread through a
string into different parts of the code base without explicitly adding
it to a context. I don't mind this if we went all in and created
better design patterns around it. In the end the amount of work
required to just get a string out of a resource container (think
IOC container) and verify it was just a little bit much.

Instead we're going to focus apiserver resources as watcher only
types. The work here is just laying the path for that.

Just hoisting the already accessible values on to the context for
now will allow any facade to get access to them.

The API facade/handle code is a freaking MESS of spaghetti code
which we should resolve at some point. Factory/vistor pattern would
do wonders here, by just adding information and routes when required.
That's for another PR.