Watcher registry
The following is the start of the watcher registry. This will manage
the lifetimes of a watcher for a give connection. Underneath it is a
basic catacomb with just some additional properties of namespacing
and counting metrics. It is somewhat backwards compatible with the
prior resources, except that we now expose real potential errors
that could happen.

There was an existing TODO to move the resources of an api handler
to a catacomb, so it makes sense to undertake this. In a future
PR we'll attach the registry directly to the apiserver, but for
now we'll duck that one.

The next task is to wire this all the way through the API server
and facade layer.