Remove state.Addresses() method
The state.Addresses() isn't used by any of the facades other than
to be aligned for an interface. Once removing that interface method
we no longer need it for APIAddresser.

Additionally, the uniter was using the state.Addresses() to decide
if we should add a controller node. EXCEPT that there will ALWAYS
be an address because of the controller api port is always added
to the state addresses so it will never be zero.