Update to log warning instead of ignoring witnesses
If any watermarks are outside of the window, then
we should force the falling behind controller to bounce and to try
again at keeping up. We don't have any mechanism to do this yet,
adding a item to table might be a waste of time. If the controller
isn't inserting into the change log, they're either bouncing all the
time or are deadlocked. If they're the latter no amount of inserting
into a table will solve the problem, as they're not reading the
change log anyway.
In addition to this, we have no theatre experience about what a
good valid window time is. For now we'll just log a warning for
visibility, before we solidify the approach.