Wire up watcher registry in place of Resources
The following starts to wire up the new watcher registry in place of
the old resources. The old resources was created prior to worker.Worker
or tomb.Tomb, so it behaves in a strange way.

The set of changes are just mechanical changes through ALL the facades.
This is a long effort, but will be rewarding if it works.