Remove change_log_witness pruning
As we won't have a controller_node table per db, then we can't
guarantee that we can clean up the witness table. Also, attempting
to keep the witness table clean for every model prune is wasteful
if the cluster is stable. The solution is to do it once when a
node is removed. That way we don't have any potential issues of
trying to run a transaction over two dbs at a critical point.