Remove apiserver/common from api package
The following removes the apiserver/common packages from the api
packages. Unfortunately, our api client sometimes brings in api
server files. When doing so, we have a tendency to bring in the
world. This can cause all sorts of problems for people who want to
use juju client in their own applications.

This is the first one I spotted. It's currently always better to
have duplication rather than a abstraction. The better way would
be to have a package that didn't tie to each layer at all (api
client and api server) and was agnostic about each side. We're not
there currently, and we have more important things to do. For now
duplication is the perfered method.