Fix data race in multiplexer
The following fixes a datarace in the multiplexer where the underlying
stream is closed and it closes the subscription channels. Any new
subscriptions then get panics about closing on a closed channel.

The solution is to expose the stream dying method, with then allows
the underlying stream to expose when it's transitioning to the dead
state. We can then prevent subscriptions from being attempted to be
unsubscribed.

In addition we prevent the unsubscribed function from being called if
we're already dying.