Cache watchable DB
The following uses the internal worker runner as a cache for
accessing the underlying watchable DB. This removes an intermidate
type as we can just expose the underlying watchable DB type.

When creating the watchable DB we already have all the pieces in
play, so we just need to align those pieces to the WatchableDB
interface. Once they're aligned we can just return that.

This means that calling `WatchableDB` is really cheap as we're not
calling `GetDB` unless we're using a new namespace.

We can do this because the underlying TxnRunner returned from
GetDB actually wraps the db, so the db can be refreshed without
refreshing the TxnRunner. If the TxnRunner fails, that will cause
everything to come down anyway.