Attempt to fix a race logic bug in dbaccessor.
The dbaccessor worker has a race condition on some logic that has the
ability to introduce a case where the dbapp can become nill after
spawning an open database worker.

The reason for this is that openDatabase can be called where the dbapp
can be in the correct state but the worker it spawns to use the dbapp
can take another x amount of time to start running and in the difference
it's possible for the dbapp to become nil from another part of the
access worker.

This commit is not trying to be the ultimate fix for this problem as the
mutex lock is not synchronised and the problem could still exist.