Allow worker retry
The following fixes the issue when rebinding the address and a
worker is starting the dbApp can be nil. To fix the problem we
check if the underlying catacomb is dying. If that's the case then
return that error. If it's not dying, force the worker to try again
with the new sentinel error.
As all errors where currently fatal, we add a check to ensure that if
the error is ErrTryAgain, that a worker can come around and attempt
to try the worker again. Upon the next time around if the catacomb
is dying, then it will correctly die. If the dbApp is still nil, it
will continue trying, otherwise it will create a worker.

We're leaning into the way the runner works here to ensure that we
can handle the case of a dbApp worker being nil. Also you can not
put the worker creation outside of the start worker function, as this
opens up possibilities of a worker being created twice and only one
of them being used.