Pass logger down to the runner
Now we're leaning on the runner implementation it would be good to
know what's happening in the runner.

The following adds a testing logger that allows the logger to be
passed to the worker runner. We now get nice logging output when
a test fails. This will improve analysis when tracking down a bug.