Fix LP1983506, local charm migration failure.
updateSeqWithMin misbehaved when providing a non 1 minimum value.
This caused models with non-sequential local charm revisions to
fail model migration. If the nextVal and minVal were the same, the
nextVal was used. However the result should have been the minVal
plus 1. The sequence saved represents the next value to use. The
minVal is the most current value available.