Add description to change log namespace id
The change_log namespace is explicitly tied to the changed key that
is being returned via the change namespace. The following updates
the change_log_namespace to add a description to ensure that if you
where checking the table rows it explained what they did.

This means that you shouldn't mix the namespace and the assoicated
key. Maybe the namespace should include the value associated with
it.