Secret watcher registry wireup
The following wires up all watchers to primarily look in the watcher
registry before looking at the resource. As we're hopefully moving
all watchers to the watcher registry so that the lifecycle of a watcher
is correctly managed. Previously, if a watcher terminates, then the
resources would still hold a reference to the dead watcher. The only
way to release it, would be to close the connection and reopen it.
The watcher registry is different in that it can let go of watchers
if they die. The watcher registry will return not found to indicate
that it's not longer avaiable.

In addition, we no longer attempt to restart the watcher upon failure.
This mirrors the old behaviour for now, but it might be worthwhile
exploring if that's correct. It might be useful to restart a watcher
if for instance there is a db error. Retrying from the client will
prevent excessive churn.