Watcher Registry Logging
Adds logging to the watcher registry so that it makes it easy to
observe what the watchers are doing per connection.

Debugging a watcher from an apiserver is actually quite difficult
if the underlying watcher dies and try to observe that from the logging
perspective.

I added a new logging label to give a wider context to this.