Generic watchers
The changes to the watcher registry only ensured that the watchers
were workers. Once bootstrapped this caused all sorts of chaos. This
is because no watcher actually aligned to the new type of watcher from
the core package.

The solution was to grab the new FirstResult internal package and shove
it one level higher. Then we can use that from the apiserver package.
From there we then needed to remove all the refrences from the state
watcher package and use the core ones. Although they're kind of similar
they are infact not quite the same in terms of behaviour. We will have
to fix that once we move everything across.
For now though, we should just force everything to be a core watcher and
be done with it.

There are virtually zero tests for the watcher package in apiserver and
considering this is the main driver of juju that is actually very
disapointing. We'll have to rectify that in the future.