Fix scshema gen output
The following starts to fix the schema gen output by forcing an
allower to check if the facade is allowed to be created before
creating the facade.

Unfortunately, we have to brute force a fake facade context through
the start of each facade. This doesn't work because of nil pointer
exceptions (panics) which can result in misleading result set.

The fix is simple, we move all the facade setup auth checks up front
in a allower check. This will prevent the construction of a context
and other mindless objects before being told that you're not allowed.

This should improve the security posture of juju and is probably worth
while doing.