Lease expiry domain
In addition to the lease domain PR (#15807) the following moves
over the expiry part of the equation to the lease domain. This keeps
everything in one location for leases.

The majority of the work was just pushing the store through the worker
scafolding and placing the sql statements in the lease domain state.

Also this cleans up the worker nicely, as we're not separating the
concerns of when it should be run and what should be run.