Scope for a given database txn runners
During an import/export it's required that we require access to
both controller and model dbs. An example of this is to export
external controllers. The external controllers are in the global
controller namespace, yet the remote applications are per model.
The solution is to provide both a controller and model at the
same time. We don't need to run this in one transaction as import/
export is managed by a 2 phase commit/saga pattern.