Import application leases during importing
The following imports the application leases during an import of
the model.

It includes wiring up the importing coordinator using the vistor
pattern. This makes it easy for unit testing as other pieces of
code do not depend on items from the domain package. Only the domain
migrations package has that knowledge.